ipq60xx: support ethernet functionality (via Qualcomm SSDK)
Ethernet can currently only be supported via the qca-nss-dp and
qca-ssdk out-of-tree kernel modules. This requires a ess-switch node,
that while ugly to look at, is reusable. The QCA-5.4 downstream
kernel declares this in the SoC dtsi. Model said approach, and declare
the common ess parameters in ipq6018-ess.dtsi.

Signed-off-by: Alexandru Gagniuc <mr.nuke.me@gmail.com>